Rename prediction to io
