+++ rampwf/utils/submission.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 import numpy as np
 import cloudpickle as pickle
 
-from .prediction import save_y_pred
+from .io import save_y_pred
 from .combine import get_score_cv_bags
 from .pretty_print import print_title, print_df_scores, print_warning
 from .scoring import score_matrix, round_df_scores, score_matrix_from_scores
